LayoutTests/http/wpt/webauthn/public-key-credential-is-user-verifying-platform-authenticator-available.html

<!DOCTYPE html>
<title>Web Authentication API: PublicKeyCredential.isUserVerifyingPlatformAuthenticatorAvailable().</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
    promise_test(t => {
        return PublicKeyCredential.isUserVerifyingPlatformAuthenticatorAvailable().then(result => {
            assert_equals(result, false);
        });
    }, "PublicKeyCredential.isUserVerifyingPlatformAuthenticatorAvailable() without any available user verifying platform authenticators.");
</script>
